+++ src/main/java/com/example/solidconnection/auth/service/SignUpService.java
@@ -3,12 +3,12 @@ package com.example.solidconnection.auth.service;
 import com.example.solidconnection.auth.dto.SignInResponse;
 import com.example.solidconnection.auth.dto.SignUpRequest;
 import com.example.solidconnection.custom.exception.CustomException;
-import com.example.solidconnection.entity.InterestedCountry;
-import com.example.solidconnection.entity.InterestedRegion;
-import com.example.solidconnection.repositories.CountryRepository;
-import com.example.solidconnection.repositories.InterestedCountyRepository;
-import com.example.solidconnection.repositories.InterestedRegionRepository;
-import com.example.solidconnection.repositories.RegionRepository;
+import com.example.solidconnection.location.country.domain.InterestedCountry;
+import com.example.solidconnection.location.region.domain.InterestedRegion;
+import com.example.solidconnection.location.country.repository.CountryRepository;
+import com.example.solidconnection.location.country.repository.InterestedCountyRepository;
+import com.example.solidconnection.location.region.repository.InterestedRegionRepository;
+import com.example.solidconnection.location.region.repository.RegionRepository;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
 import org.springframework.transaction.annotation.Transactional;
