+++ src/main/java/com/example/solidconnection/location/country/domain/Country.java
@@ -1,5 +1,6 @@
-package com.example.solidconnection.entity;
+package com.example.solidconnection.location.country.domain;
 
+import com.example.solidconnection.location.region.domain.Region;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
