+++ src/main/java/com/example/solidconnection/location/country/domain/InterestedCountry.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.entity;
+package com.example.solidconnection.location.country.domain;
 
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import jakarta.persistence.Entity;
