+++ src/main/java/com/example/solidconnection/location/country/repository/CountryRepository.java
@@ -1,6 +1,6 @@
-package com.example.solidconnection.repositories;
+package com.example.solidconnection.location.country.repository;
 
-import com.example.solidconnection.entity.Country;
+import com.example.solidconnection.location.country.domain.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
