+++ src/main/java/com/example/solidconnection/location/country/repository/InterestedCountyRepository.java
@@ -1,6 +1,6 @@
-package com.example.solidconnection.repositories;
+package com.example.solidconnection.location.country.repository;
 
-import com.example.solidconnection.entity.InterestedCountry;
+import com.example.solidconnection.location.country.domain.InterestedCountry;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
