+++ src/main/java/com/example/solidconnection/university/domain/University.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.university.domain;
 
-import com.example.solidconnection.entity.Country;
-import com.example.solidconnection.entity.Region;
+import com.example.solidconnection.location.country.domain.Country;
+import com.example.solidconnection.location.region.domain.Region;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
