+++ src/test/java/com/example/solidconnection/location/country/fixture/CountryFixtureBuilder.java
@@ -1,8 +1,9 @@
-package com.example.solidconnection.country.fixture;
+package com.example.solidconnection.location.country.fixture;
 
-import com.example.solidconnection.entity.Country;
-import com.example.solidconnection.entity.Region;
-import com.example.solidconnection.country.repository.CountryRepositoryForTest;
+
+import com.example.solidconnection.location.country.repository.CountryRepositoryForTest;
+import com.example.solidconnection.location.country.domain.Country;
+import com.example.solidconnection.location.region.domain.Region;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.test.context.TestComponent;
