+++ src/test/java/com/example/solidconnection/location/country/repository/CountryRepositoryForTest.java
@@ -1,6 +1,6 @@
-package com.example.solidconnection.country.repository;
+package com.example.solidconnection.location.country.repository;
 
-import com.example.solidconnection.entity.Country;
+import com.example.solidconnection.location.country.domain.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
