+++ src/test/java/com/example/solidconnection/location/region/repository/RegionRepositoryForTest.java
@@ -1,6 +1,6 @@
-package com.example.solidconnection.region.repository;
+package com.example.solidconnection.location.region.repository;
 
-import com.example.solidconnection.entity.Region;
+import com.example.solidconnection.location.region.domain.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
