+++ src/test/java/com/example/solidconnection/support/integration/BaseIntegrationTest.java
@@ -6,14 +6,14 @@ import com.example.solidconnection.application.domain.LanguageTest;
 import com.example.solidconnection.application.repository.ApplicationRepository;
 import com.example.solidconnection.community.board.domain.Board;
 import com.example.solidconnection.community.board.repository.BoardRepository;
-import com.example.solidconnection.entity.Country;
+import com.example.solidconnection.location.country.domain.Country;
 import com.example.solidconnection.community.post.domain.PostImage;
-import com.example.solidconnection.entity.Region;
+import com.example.solidconnection.location.region.domain.Region;
 import com.example.solidconnection.community.post.domain.Post;
 import com.example.solidconnection.community.post.repository.PostRepository;
-import com.example.solidconnection.repositories.CountryRepository;
+import com.example.solidconnection.location.country.repository.CountryRepository;
 import com.example.solidconnection.community.post.repository.PostImageRepository;
-import com.example.solidconnection.repositories.RegionRepository;
+import com.example.solidconnection.location.region.repository.RegionRepository;
 import com.example.solidconnection.score.domain.GpaScore;
 import com.example.solidconnection.score.domain.LanguageTestScore;
 import com.example.solidconnection.score.repository.GpaScoreRepository;
