+++ src/test/java/com/example/solidconnection/university/fixture/UniversityFixture.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.university.fixture;
 
-import com.example.solidconnection.country.fixture.CountryFixture;
-import com.example.solidconnection.region.fixture.RegionFixture;
+import com.example.solidconnection.location.country.fixture.CountryFixture;
+import com.example.solidconnection.location.region.fixture.RegionFixture;
 import com.example.solidconnection.university.domain.University;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.test.context.TestComponent;
