+++ src/test/java/com/example/solidconnection/university/service/UniversityRecommendServiceTest.java
@@ -1,9 +1,9 @@
 package com.example.solidconnection.university.service;
 
-import com.example.solidconnection.country.fixture.CountryFixture;
+import com.example.solidconnection.location.country.fixture.CountryFixture;
 import com.example.solidconnection.entity.InterestedCountry;
 import com.example.solidconnection.entity.InterestedRegion;
-import com.example.solidconnection.region.fixture.RegionFixture;
+import com.example.solidconnection.location.region.fixture.RegionFixture;
 import com.example.solidconnection.repositories.InterestedCountyRepository;
 import com.example.solidconnection.repositories.InterestedRegionRepository;
 import com.example.solidconnection.siteuser.domain.SiteUser;
